back/requirements.txt
asgiref==3.7.2
Django==4.0.6
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
Pillow==9.2.0
PyJWT==1.7.1
pytz==2022.1
sqlparse==0.4.4
tzdata==2023.3
